requirements.txt
torch
torchvision
datasets
tqdm
pillow
wandb
pynvml
deepspeed
